feat: Add organization email verification popup and settings management
- Implemented OrgVerificationPopup component for email verification.
- Integrated email verification process in OrganizationSettings component.
- Added image upload functionality in ProfileSettings and OrganizationSettings.
- Created reusable ConfirmationDialog component for user confirmations.
- Enhanced form validation for organization and profile settings.
- Updated styles for new components and improved user experience.
