+++ src/components/Settings/Account/VerificationPopup.jsx
@@ -1,28 +1,19 @@
-
 import React, { useState } from 'react';
 import axios from 'axios';
 import { toast } from 'react-toastify';
 import { Dialog } from '@mui/material';
 import { Shield, Mail, Check, X } from 'lucide-react';
-import { getAuthHeader } from '../../utils/authHeader';
-import { decodeToken, useUser } from '../../contexts/UserContext';
-import { BASE_URLS } from '../../services/api/config';
-import { useThemeStyles } from '../../hooks/useThemeStyles';
-import './Styles/VerifyPopup.css';
+import { getAuthHeader } from '../../../utils/authHeader';
+import { decodeToken, useUser } from '../../../contexts/UserContext';
+import { BASE_URLS } from '../../../services/api/config';
+import { useThemeStyles } from '../../../hooks/useThemeStyles';
+import '../Styles/VerifyPopup.css';
 
-function OrgVerificationPopup({ onClose, email, code }) {
+function VerificationPopup({ onClose, email, code }) {
   // Local state to hold the user-entered verification code
   const [inputcode, setInputCode] = useState('');
   const { userData } = useUser();
   const { updateCSSVariables } = useThemeStyles();
-  // Get orgId from user context or decoded token
-  let orgId = null;
-  if (userData?.org_id) {
-    orgId = userData.org_id;
-  } else {
-    const decoded = decodeToken();
-    orgId = decoded?.org_id;
-  }
 
   // Apply theme variables when component mounts
   React.useEffect(() => {
@@ -43,11 +34,10 @@ function OrgVerificationPopup({ onClose, email, code }) {
           userId = decoded?.id;
         }
         if (!userId) throw new Error('User ID not found in token');
-        if (!orgId) throw new Error('Organization ID not found');
 
         // Send verified email update to the server with Authorization header
         await axios.put(
-          `${BASE_URLS.orgsettings}/email/${orgId}`,
+          `${BASE_URLS.settings}/email/${userId}`,
           { email },
           {
             headers: {
@@ -56,39 +46,15 @@ function OrgVerificationPopup({ onClose, email, code }) {
           }
         );
 
-        toast.success('Verification successful!', {
-          position: "top-center",
-          autoClose: 3000,
-          hideProgressBar: false,
-          closeOnClick: true,
-          pauseOnHover: true,
-          draggable: true,
-          style: { zIndex: 9999 }
-        });
+        toast.success('Verification successful!');
         onClose(); // Close the popup after successful verification
       } catch (error) {
         console.error('There was an error verifying the email!', error);
-        toast.error('Verification failed. Please try again.', {
-          position: "top-center",
-          autoClose: 3000,
-          hideProgressBar: false,
-          closeOnClick: true,
-          pauseOnHover: true,
-          draggable: true,
-          style: { zIndex: 9999 }
-        });
+        toast.error('Verification failed. Please try again.');
       }
     } else {
       // Show error if code does not match
-      toast.error('Invalid verification code. Please try again.', {
-        position: "top-center",
-        autoClose: 3000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        style: { zIndex: 9999 }
-      });
+      toast.error('Invalid verification code. Please try again.');
     }
   };
 
@@ -98,8 +64,8 @@ function OrgVerificationPopup({ onClose, email, code }) {
       onClose={onClose}
       maxWidth="sm"
       fullWidth
-      aria-labelledby="org-verification-popup-title"
-      aria-describedby="org-verification-popup-description"
+      aria-labelledby="verification-popup-title"
+      aria-describedby="verification-popup-description"
       BackdropProps={{
         className: 'verify-popup-backdrop'
       }}
@@ -115,8 +81,8 @@ function OrgVerificationPopup({ onClose, email, code }) {
           <div className="verify-icon">
             <Shield className="shield-icon" />
           </div>
-          <h1 className="verify-title">Organization Email Verification</h1>
-          <p className="verify-subtitle">Verify your organization email address</p>
+          <h1 className="verify-title">Email Verification</h1>
+          <p className="verify-subtitle">Verify your email address to secure your account</p>
         </div>
         
         <form className="verify-form" onSubmit={handleSubmit}>
@@ -164,4 +130,4 @@ function OrgVerificationPopup({ onClose, email, code }) {
   );
 }
 
-export default OrgVerificationPopup;
+export default VerificationPopup;
