+++ src/pages/Shared/Settings.jsx
@@ -1,13 +1,13 @@
 
 import React, { useState } from 'react';
-import ProfileSection from './../../components/Settings/ProfileSettings';
-import AccountSection from './../../components/Settings/AccountSettings';
-import OrganizationSection from './../../components/Settings/OrganizationSettings';
+import ProfileSection from './../../components/Settings/Profile/ProfileSettings';
+import AccountSection from './../../components/Settings/Account/AccountSettings';
+import OrganizationSection from './../../components/Settings/Organization/OrganizationSettings';
 import AdminLayout from './../../layouts/Admin/AdminLayout';
 import UserLayout from './../../layouts/User/UserLayout';
 import { Tabs, Tab, Box, Paper, Typography, Card, CardContent, Stack } from '@mui/material';
 import { useUser } from './../../contexts/UserContext';
-import AppearanceSettings from './../../components/Settings/AppearanceSettings';
+import AppearanceSettings from './../../components/Settings/Appearance/AppearanceSettings';
 import { Navigate } from 'react-router-dom';
 
 const Settings = () => {
